Move common files to folder.
