+++ packages/style/src/common/sheet.ts
@@ -1,5 +1,5 @@
 import { isAtRule, isImportRule, insertAtRule, insertImportRule, insertRule } from './helpers';
-import { SheetMap, SheetManager } from './types';
+import { SheetMap, SheetManager } from '../types';
 
 export function createSheetManager(sheets: SheetMap): SheetManager {
   return {
