+++ packages/style/src/common/syntax.ts
