+++ packages/style/src/index.ts
@@ -3,10 +3,10 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
-export * from './cache';
-export * from './constants';
-export * from './engine';
-export * from './helpers';
-export * from './sheet';
-export * from './syntax';
+export * from './common/cache';
+export * from './common/constants';
+export * from './common/engine';
+export * from './common/helpers';
+export * from './common/sheet';
+export * from './common/syntax';
 export * from './types';
